# docs/swaggerfile.yml
openapi: 3.0.1
info:
  title: FinTrack.Web
  version: '1.0'
servers:
  - url: /
paths:
  /register:
    post:
      tags:
        - Authentication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '200':
          description: OK
  /login:
    post:
      tags:
        - Authentication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/LoginUserRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/LoginUserRequest'
      responses:
        '200':
          description: OK
  /countries:
    post:
      tags:
        - Countries
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCountryRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateCountryRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateCountryRequest'
      responses:
        '200':
          description: OK
    get:
      tags:
        - Countries
      parameters:
        - name: search_query
          in: query
          schema:
            type: string
            default: ''
        - name: page
          in: query
          schema:
            type: integer
            format: int32
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: OK
  '/countries/{id}':
    get:
      tags:
        - Countries
      operationId: GetCountryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/GetSecurityResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/GetSecurityResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/GetSecurityResponse'
  /currencies:
    post:
      tags:
        - Currencies
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCurrencyRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateCurrencyRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateCurrencyRequest'
      responses:
        '200':
          description: OK
    get:
      tags:
        - Currencies
      parameters:
        - name: search_query
          in: query
          schema:
            type: string
            default: ''
        - name: page
          in: query
          schema:
            type: integer
            format: int32
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: OK
  '/currencies/{id}':
    get:
      tags:
        - Currencies
      operationId: GetCurrencyById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
  '/operations/{operationId}':
    delete:
      tags:
        - Operations
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
  /securities:
    post:
      tags:
        - Securities
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSecurityRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateSecurityRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateSecurityRequest'
      responses:
        '200':
          description: OK
  '/securities/{securityId}':
    get:
      tags:
        - Securities
      operationId: GetSecurityById
      parameters:
        - name: securityId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
    patch:
      tags:
        - Securities
      parameters:
        - name: securityId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSecurityRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateSecurityRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateSecurityRequest'
      responses:
        '200':
          description: OK
  '/securities/{securityId}/operations':
    get:
      tags:
        - Securities
      parameters:
        - name: securityId
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: page
          in: query
          schema:
            type: integer
            format: int32
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: OK
  '/securities/{securityId}/status':
    get:
      tags:
        - Securities
      parameters:
        - name: securityId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
components:
  schemas:
    Country:
      required:
        - alpha2Code
        - alpha3Code
        - name
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        alpha2Code:
          type: string
          nullable: true
        alpha3Code:
          type: string
          nullable: true
      additionalProperties: false
    CreateCountryRequest:
      required:
        - name
      type: object
      properties:
        name:
          minLength: 1
          type: string
        alpha2Code:
          maxLength: 2
          minLength: 2
          type: string
          nullable: true
        alpha3Code:
          maxLength: 3
          minLength: 3
          type: string
          nullable: true
      additionalProperties: false
    CreateCurrencyRequest:
      required:
        - name
      type: object
      properties:
        name:
          maxLength: 100
          minLength: 1
          type: string
        alpha3Code:
          maxLength: 3
          minLength: 3
          type: string
          nullable: true
        symbol:
          maxLength: 10
          type: string
          nullable: true
        decimals:
          maximum: 65535
          minimum: 0
          type: integer
          format: int32
          nullable: true
        numberToMajor:
          maximum: 65535
          minimum: 0
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
    CreateOperationRequest:
      type: object
      properties:
        operationType:
          $ref: '#/components/schemas/OperationType'
        operationDate:
          type: string
          format: date
        value:
          minimum: 0
          exclusiveMinimum: true
          type: number
          format: double
        quantity:
          minimum: 0
          exclusiveMinimum: true
          type: integer
          format: int32
        foreignTaxes:
          minimum: 0
          type: number
          format: double
        expensesAndCharges:
          minimum: 0
          type: number
          format: double
      additionalProperties: false
    CreateSecurityRequest:
      required:
        - name
        - nativeCurrency
        - sourceCountry
      type: object
      properties:
        name:
          minLength: 1
          type: string
        isin:
          type: string
          nullable: true
        nativeCurrency:
          type: integer
          format: int32
        counterpartyCountry:
          type: integer
          format: int32
          nullable: true
        sourceCountry:
          type: integer
          format: int32
        issuingNIF:
          type: string
          nullable: true
        operations:
          type: array
          items:
            $ref: '#/components/schemas/CreateOperationRequest'
          nullable: true
      additionalProperties: false
    CreateUser:
      required:
        - email
        - password
        - userName
      type: object
      properties:
        userName:
          minLength: 1
          type: string
        email:
          minLength: 1
          type: string
          format: email
        password:
          minLength: 1
          type: string
      additionalProperties: false
    Currency:
      required:
        - alpha3Code
        - name
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        alpha3Code:
          type: string
          nullable: true
        symbol:
          type: string
          nullable: true
        decimals:
          type: integer
          format: int32
        numberToMajor:
          type: integer
          format: int32
      additionalProperties: false
    GetSecurityResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          nullable: true
        isin:
          type: string
          nullable: true
        nativeCurrency:
          $ref: '#/components/schemas/Currency'
        counterpartyCountry:
          $ref: '#/components/schemas/Country'
        sourceCountry:
          $ref: '#/components/schemas/Country'
        issuingNIF:
          type: string
          nullable: true
      additionalProperties: false
    LoginUserRequest:
      type: object
      properties:
        identity:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        rememberMe:
          type: boolean
      additionalProperties: false
    OperationType:
      enum:
        - Sell
        - Buy
      type: string